\name{dynaTree-class}
\Rdversion{1.1}
\docType{class}
\alias{dynaTree-class}
\alias{retire-methods}
\alias{copy-methods}
\alias{sens-methods}
\alias{alcX-methods}
\alias{alc-methods}
\alias{entropyX-methods}
\alias{ieci-methods}
\alias{classprobs-methods}
\alias{rejuvenate-methods}
\alias{relevance-methods}
\alias{intervals-methods}
\alias{retire,dynaTree-method}
\alias{sens,dynaTree-method}
\alias{copy,dynaTree-method}
\alias{alcX,dynaTree-method}
\alias{alc,dynaTree-method}
\alias{entropyX,dynaTree-method}
\alias{ieci,dynaTree-method}
\alias{classprobs,dynaTree-method}
\alias{rejuvenate,dynaTree-method}
\alias{relevance,dynaTree-method}
\alias{intervals,dynaTree-method}
\alias{sens}
\alias{retire}
\alias{copy}
\alias{alcX}
\alias{alc}
\alias{entropyX}
\alias{ieci}
\alias{classprobs}
\alias{rejuvenate}
\alias{relevance}
\alias{intervals}
\title{Class "dynaTree"}
\description{
	A stub for class dynaTree and its custom
	generic methods
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{retire}{\code{signature(object = "dynaTree")}: ... }
    \item{sens}{\code{signature(object = "dynaTree")}: ... }
	 }
}
\details{
 This is just a stub file.  See \code{\link{sens.dynaTree}}
 and \code{\link{retire.dynaTree}} for more information on
 the generic methods used in this package
} 
\references{

 Taddy, M.A., Gramacy, R.B., and Polson, N. (2011).
  \dQuote{Dynamic trees for learning and design}
  Journal of the American Statistical Association, 106(493), pp. 109-123;
  arXiv:0912.1586

  Gramacy, R.B., Taddy, M.A., and S. Wild (2011).
 \dQuote{Variable Selection and Sensitivity Analysis via
   Dynamic Trees with an Application to Computer Code Performance Tuning}
  arXiv:1108.4739
  
  \url{http://faculty.chicagobooth.edu/robert.gramacy/dynaTree.html}

}
\author{
Robert B. Gramacy \email{rbgramacy@chicagobooth.edu} and\cr
Matt Taddy \email{taddy@chicagobooth.edu}}

\seealso{
	\code{\link{dynaTree}}, \code{\link{predict.dynaTree}}, 
	\code{\link{update.dynaTree}}, \code{\link{retire.dynaTree}},
	\code{\link{sens.dynaTree}}, \code{\link{alcX.dynaTree}}
}
\examples{
showClass("dynaTree")
}
\keyword{classes}
\keyword{methods}
