\name{dynaTree-package}
\alias{dynaTree-package}
\docType{package}
\title{Dynamic trees for learning and design}
\description{Inference by sequential Monte Carlo for
  dynamic tree regression and classification models 
  with hooks provided for sequential design and optimization, 
  fully online learning with drift, variable selection, and   
  sensitivity analysis of inputs.  Illustrative
  examples from the original dynamic trees paper are facilitated
  by demos in the package; see demo(package="dynaTree")
}
\details{
For a fuller overview including a complete list of functions, and
demos, please use \code{help(package="dynaTree")}.
}
\author{
Robert B. Gramacy \email{rbgramacy@chicagobooth.edu} and\cr
Matt Taddy \email{taddy@chicagobooth.edu}
}
\references{
  
 Taddy, M.A., Gramacy, R.B., and Polson, N. (2011).
  \dQuote{Dynamic trees for learning and design}
  Journal of the American Statistical Association, 106(493), pp. 109-123;
  arXiv:0912.1586

  \url{http://faculty.chicagobooth.edu/robert.gramacy/dynaTree.html}
}
\keyword{ package }
\seealso{ \pkg{plgp}, \pkg{tgp} }
