\name{dynaTree-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{plotprogress}
\alias{plotal}
\alias{alcalc}
%%% THINGS THAT DO NEED TO BE DOCUMENTED EVENTUALLY
\alias{classprobs.dynaTree}
\alias{ieci.dynaTree}
\alias{intervals.dynaTree}
\alias{sameleaf.dynaTree}
\alias{qEI.dynaTree}
\alias{qEI}
\alias{qEntropy.dynaTree}
\alias{qEntropy}
\title{Internal dynaTree Functions}
\description{
  Internal \pkg{dynaTree} functions
}
\details{
  These functions are primarily provided to facilitate the
  demos, and to provide for a cleaner exposition.  We encourage
  users to inspect their contents in order to help develop 
  similar functions to suit their particular needs
}
\keyword{ internal }
