\name{dynaTree-package}
\alias{dynaTree-package}
\docType{package}
\title{Dynamic trees for learning and design}
\description{Inference by sequential Monte Carlo for 
  dynamic tree regression and classification models by
  particle learning (PL), with hooks provided for sequential 
  design and optimization.  Illustrative examples
  from the original dynamic trees paper are facilitated
  by demos in the package; see demo(package="dynaTree")}
\details{
For a fuller overview including a complete list of functions, and
demos, please use \code{help(package="dynaTree")}.
}
\author{
Robert B. Gramacy \email{bobby@statslab.cam.ac.uk} and\cr
Matt Taddy \email{taddy@chicagobooth.edu}
}
\references{
  
  Taddy, M.A., Gramacy, R.B., and Polson, N. (2009).
  \dQuote{Dynamic trees for learning and design}
  Tech. Rep. arXiv:0912.1636, University of Cambridge.

  \url{http://www.statslab.cam.ac.uk/~bobby/dynaTree.html}
}
\keyword{ package }
\seealso{ \pkg{plgp}, \pkg{tgp} }
